backend/requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
scikit-learn
python-dotenv
httpx
google-generativeai
